+++ deno.json
@@ -1,6 +1,6 @@
 {
   "tasks": {
-    "dev": "deno run -A lib/watch.ts html components lib sitemap assets src -- deno run --lock-write -A main.ts"
+    "dev": "deno run -A lib/watch.ts components lib assets routes plugins -- deno run --lock-write -A main.tsx"
   },
   "lint": {
     "rules": {
@@ -12,13 +12,13 @@
     "exclude": ["legacy"]
   },
   "compilerOptions": {
+    "lib": ["deno.ns", "dom", "dom.iterable"],
     "jsx": "react-jsx",
-    "jsxImportSource": "hastx"
+    "jsxImportSource": "revolution"
   },
   "imports": {
-    "effection": "https://deno.land/x/effection@3.0.0-alpha.9/mod.ts",
-    "freejack/": "./lib/",
-    "hastx/": "https://deno.land/x/hastx@v0.0.6/",
-    "hastx/jsx-runtime": "https://deno.land/x/hastx@v0.0.6/jsx-runtime.ts"
+    "effection": "https://deno.land/x/effection@3.0.0-beta.2/mod.ts",
+    "revolution": "https://deno.land/x/revolution@0.3.1/mod.ts",
+    "revolution/jsx-runtime": "https://deno.land/x/revolution@0.3.1/jsx-runtime.ts"
   }
 }
