--- html.ts
@@ -1,5 +1,0 @@
-import AppHtml from "./html/app.html.tsx";
-import IndexHtml from "./html/index.html.tsx";
-export * from "./html/advance-backstage-plugin-development.html.tsx";
-
-export { AppHtml, IndexHtml };
