--- lib/deps.ts
@@ -1,5 +1,0 @@
-export {
-  match,
-  type MatchFunction,
-  type MatchResult,
-} from "https://deno.land/x/path_to_regexp@v6.2.1/index.ts";
