--- lib/router.ts
@@ -1,1 +1,0 @@
-export const main = Symbol.for("main");
