+++ main.ts
@@ -1,16 +1,1 @@
-import { main, suspend } from "effection";
-
-import { useServer } from "freejack/server.ts";
-import serve from "./sitemap/sitemap.ts";
-
-await main(function* () {
-  let server = yield* useServer({
-    serve,
-    port: 8000,
-    dir: new URL(".", import.meta.url).pathname,
-  });
-
-  console.log(`www -> http://${server.hostname}:${server.port}`);
-
-  yield* suspend();
-});
+import "./main.tsx";
